Get api tests running locally

+++ apitests/run-api-tests.sh
@@ -5,7 +5,9 @@ phpServerId=$!
 
 # Run the tests
 tavern-ci ./test_errors.tavern.yaml
+if [ $? != 0 ]; then exit 1; fi
 tavern-ci ./test_use_case.tavern.yaml
+if [ $? != 0 ]; then exit 1; fi
 
 # Stop the server
 kill -3 $phpServerId
